+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 #### Changed
 - Test now raises a `OpenapiFirst::Test::UnknownQueryParameterError` when it sees unknown query parameters
+- Test does not track requests/responses unless the OAD was registered via Test.register (or OpenapiFirst.register)
 
 ## 3.0.1
 - Add missing gem dependency "drb", which is no longer installed by default with newer rubies. This is used in openapi_first/test to make parallel tests work.